Extract AuthButtons
